fix: ajusta tipagem e requests comentados
