+++ pages/admin.tsx
@@ -1,3 +1,4 @@
+import React from "react";
 import { Admin } from "../components/Admin";
 
 const admin = () => <Admin />;
